# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/item-providers": {
      "post": {
        "description": "Admin creates an item provider linked to a user account. The user referenced by userId must have type provider.",
        "operationId": "ItemProvidersController_createItemProvider_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemProviderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item provider created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemProviderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "User is not type provider, account mismatch, or invalid input"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create item provider",
        "tags": [
          "Item Providers"
        ]
      }
    },
...
    "/api/v1/accounts": {
      "post": {
        "description": "Creates an NFC-backed account for a user. Requires an admin API token.",
        "operationId": "AccountsController_createAccount_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAccountResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create account",
        "tags": [
          "Accounts"
        ]
      },
      "get": {
        "description": "Returns a paginated list of accounts with each account owner (user) summary. Admins may filter by owner user id, owner user type, lock state, and a case-insensitive search on NFC UID or owner name fields.",
        "operationId": "AccountsController_listAccounts_v1",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Return only accounts owned by this user.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "ownerType",
            "required": false,
            "in": "query",
            "description": "Return only accounts whose owner has this user type.",
            "schema": {
              "type": "string",
              "enum": [
                "user",
                "provider",
                "admin"
              ]
            }
          },
          {
            "name": "isLocked",
            "required": false,
            "in": "query",
            "description": "Return only locked (true) or unlocked (false) accounts.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive substring match on NFC card UID, owner display name, or owner login name (userName).",
            "schema": {
              "maxLength": 128,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (default 50, max 100).",
            "schema": {
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of rows to skip (pagination).",
            "schema": {
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Accounts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAccountsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List accounts",
        "tags": [
          "Accounts"
        ]
      }
    },
...
    "/api/v1/admin/mint": {
      "post": {
        "description": "Increases an account balance and records a mint transaction on the chain.",
        "operationId": "AdminController_mint_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MintDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mint applied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MintResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mint currency",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/item-providers": {
      "get": {
        "description": "Returns a paginated list of item providers with owner and linked NFC metadata. Admins may filter by user id, owner type, and search.",
        "operationId": "AdminController_listItemProviders_v1",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Return only providers linked to this user id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "ownerType",
            "required": false,
            "in": "query",
            "description": "Return only providers whose owner has this user type.",
            "schema": {
              "type": "string",
              "enum": [
                "user",
                "provider",
                "admin"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive substring match on provider name, owner display name, owner login name, or linked account NFC UID.",
            "schema": {
              "maxLength": 128,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (default 50, max 100).",
            "schema": {
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of rows to skip (pagination).",
            "schema": {
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item providers retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListItemProvidersResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List item providers",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/item-providers/{id}": {
      "patch": {
        "description": "Updates display name and/or the linked user and account. When relinking, the user must have type provider and the account must belong to that user.",
        "operationId": "AdminController_updateItemProvider_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Item provider ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemProviderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item provider updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemProviderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or relink constraints"
          },
          "404": {
            "description": "Item provider not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update item provider",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/accounts": {
      "get": {
        "description": "Returns a paginated list of accounts with each account owner (user) summary. Same behavior as GET /accounts.",
        "operationId": "AdminController_listAccounts_v1",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Return only accounts owned by this user.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "ownerType",
            "required": false,
            "in": "query",
            "description": "Return only accounts whose owner has this user type.",
            "schema": {
              "type": "string",
              "enum": [
                "user",
                "provider",
                "admin"
              ]
            }
          },
          {
            "name": "isLocked",
            "required": false,
            "in": "query",
            "description": "Return only locked (true) or unlocked (false) accounts.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive substring match on NFC card UID, owner display name, or owner login name (userName).",
            "schema": {
              "maxLength": 128,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (default 50, max 100).",
            "schema": {
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of rows to skip (pagination).",
            "schema": {
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Accounts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAccountsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List accounts",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/accounts/{accountId}": {
      "patch": {
        "description": "Updates admin-managed fields such as lock state.",
        "operationId": "AdminController_updateAccount_v1",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "description": "Account ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAccountPatchResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "No fields to update"
          },
          "404": {
            "description": "Account not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update account (admin)",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Classurency API",
    "description": "API documentation for the Classurency backend.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "token",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateItemProviderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Main shop counter"
          },
          "userId": {
            "type": "string",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "description": "The user that owns this provider. Must have user type provider."
          },
          "accountId": {
            "type": "string",
            "example": "9c7bb5eb-1684-49b2-bf1d-7f8ec8f0b6a4",
            "description": "Account to receive funds for this provider. Must belong to userId."
          }
        },
        "required": [
          "name",
          "userId",
          "accountId"
        ]
      },
      "ItemProviderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Item-provider ID",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Display name"
          },
          "userId": {
            "type": "string",
            "description": "Owning user ID",
            "format": "uuid"
          },
          "accountId": {
            "type": "string",
            "description": "Linked account ID",
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "name",
          "userId",
          "accountId",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateAccountDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "format": "uuid",
            "description": "User id."
          }
        },
        "required": [
          "userId"
        ]
      },
      "CreateAccountResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "name": {
            "type": "string",
            "description": "User display name"
          },
          "nfcCardUid": {
            "type": "string",
            "description": "NFC card UID assigned to the account"
          },
          "publicKeyHex": {
            "type": "string",
            "description": "Account public key (hex)"
          },
          "pin": {
            "type": "string",
            "description": "4-digit PIN for the NFC card"
          },
          "encryptedPrivateKeyForCard": {
            "type": "string",
            "description": "Encrypted private key for programming onto the card (one-time delivery)."
          }
        },
        "required": [
          "id",
          "name",
          "nfcCardUid",
          "publicKeyHex",
          "pin",
          "encryptedPrivateKeyForCard"
        ]
      },
      "AccountListOwnerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "user",
              "provider",
              "admin"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "userName",
          "type",
          "createdAt",
          "updatedAt"
        ]
      },
      "ListAccountItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "nfcCardUid": {
            "type": "object",
            "nullable": true,
            "description": "NFC card UID when assigned."
          },
          "publicKeyHex": {
            "type": "string"
          },
          "balance": {
            "type": "number",
            "description": "Account balance."
          },
          "isLocked": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "owner": {
            "description": "The user that owns this account.",
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountListOwnerDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "userId",
          "publicKeyHex",
          "balance",
          "isLocked",
          "createdAt",
          "updatedAt",
          "owner"
        ]
      },
      "ListAccountsResponseDto": {
        "type": "object",
        "properties": {
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListAccountItemDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total rows matching the filters (ignoring limit/offset)."
          }
        },
        "required": [
          "accounts",
          "total"
        ]
      },
...
      "MintDto": {
        "type": "object",
        "properties": {
          "adminUserAccountId": {
            "type": "string",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "format": "uuid",
            "description": "Admin user account id."
          },
          "accountId": {
            "type": "string",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "format": "uuid",
            "description": "Target account id."
          },
          "amount": {
            "type": "number",
            "example": 10,
            "minimum": 0.01,
            "description": "Amount to mint."
          },
          "description": {
            "type": "string",
            "example": "Initial top-up",
            "maxLength": 500,
            "description": "Optional mint description."
          }
        },
        "required": [
          "adminUserAccountId",
          "accountId",
          "amount"
        ]
      },
      "MintResponseDto": {
        "type": "object",
        "properties": {
          "balance": {
            "type": "string",
            "example": "100.00",
            "description": "Account balance after mint (decimal string)."
          }
        },
        "required": [
          "balance"
        ]
      },
      "ListItemProviderItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "linkedAccountNfcCardUid": {
            "type": "object",
            "nullable": true,
            "description": "NFC card UID of the linked account when assigned."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "owner": {
            "description": "The user that owns this item provider.",
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountListOwnerDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "userId",
          "accountId",
          "createdAt",
          "updatedAt",
          "owner"
        ]
      },
      "ListItemProvidersResponseDto": {
        "type": "object",
        "properties": {
          "itemProviders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListItemProviderItemDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total rows matching the filters (ignoring limit/offset)."
          }
        },
        "required": [
          "itemProviders",
          "total"
        ]
      },
      "UpdateItemProviderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name",
            "minLength": 1,
            "maxLength": 128
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "Owning user id (must be sent together with accountId)."
          },
          "accountId": {
            "type": "string",
            "format": "uuid",
            "description": "Linked account id (must be sent together with userId)."
          }
        }
      },
      "UpdateAccountAdminDto": {
        "type": "object",
        "properties": {
          "isLocked": {
            "type": "boolean",
            "description": "When true, the account cannot be used for login or transfers until unlocked."
          }
        }
      },
      "AdminAccountPatchResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "isLocked": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "isLocked",
          "updatedAt"
        ]
      },
...
    }
  }
}